Create Alarm clock
This function takes an alarm time in the format of a datetime.time object, and will continuously check the current time until it matches the alarm time. When the alarm time is reached, the function will print "Wake up! It's [alarm_time]" and return. It waits for one second between each check of the current time.
You can use this function by passing the alarm time in the format of datetime.time(hour=8, minute=30) and call the function alarm_clock(alarm_time)